+++ notUsed/Weapon11.java
@@ -10,7 +10,7 @@ public class Weapon11 extends NormalWeapon{
     public Weapon11(Entity owner) {
         super(owner);
     }
-    public void update(float time){
+    public void update(double time){
         super.update(time);
     }
     public void shoot(){
@@ -18,9 +18,9 @@ public class Weapon11 extends NormalWeapon{
             Vec2d input=owner.getTargetingPos();
             Vec2d pos=owner.getBulletPosition();
             if (input==null) return;
-            float size=this.size*1.8;
-            float health=this.health*20;
-            float damage=this.damage*0.8;
+            double size=this.size*1.8;
+            double health=this.health*20;
+            double damage=this.damage*0.8;
             Vec2d velocity=input.limit(speed*4);
             for(int i=-3;i<=3;i++) {
                 AimBullet b = shootAimBullet(pos.add(velocity), velocity.rotate(i*30).multiply(1.5), size, health, damage);
