+++ notUsed/Weapon7.java
@@ -18,9 +18,9 @@ public class Weapon7 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
             Vec2d velocity=input.limit(speed*9).add(Util.randomVec().limit(0.001));
-            float size=this.size*0.5;
-            float health=this.health*20;
-            float damage=this.damage*5;
+            double size=this.size*0.5;
+            double health=this.health*20;
+            double damage=this.damage*5;
             shootBullet(pos,velocity,size,health,damage);
             cooldown= round(reload*5);
         }
